tweak for ESP-IDF v5

+++ src/lgfx/v1_autodetect/LGFX_AutoDetect_ESP32.hpp
@@ -442,7 +442,7 @@ namespace lgfx
       bus->endTransaction();
       _pin_level(pin_cs, true);
 
-      ESP_LOGW(LIBRARY_NAME, "[Autodetect] read cmd:%02x = %08x", cmd, res);
+      ESP_LOGW(LIBRARY_NAME, "[Autodetect] read cmd:%02x = %08x", (unsigned int)cmd, (unsigned int)res);
       return res;
     }
 
@@ -474,7 +474,7 @@ namespace lgfx
       {
         nvs_get_u32(nvs_handle, NVS_KEY, static_cast<uint32_t*>(&nvs_board));
         nvs_close(nvs_handle);
-        ESP_LOGW(LIBRARY_NAME, "[Autodetect] load from NVS : board:%d", nvs_board);
+        ESP_LOGW(LIBRARY_NAME, "[Autodetect] load from NVS : board:%d", (int)nvs_board);
       }
 
       if (0 == nvs_board)
@@ -1348,7 +1348,7 @@ namespace lgfx
               _bus_spi.endTransaction();
               lgfx::gpio_hi(GPIO_NUM_15);
               id = buf[0] << 24 | buf[1] << 16 | buf[2] << 8 | buf[3];
-              ESP_LOGW(LIBRARY_NAME, "[Autodetect] panel size :%08x", id);
+              ESP_LOGW(LIBRARY_NAME, "[Autodetect] panel size :%08x", (unsigned int)id);
               if (id == 0x03C0021C)
               {  //  check panel ( panel size 960(0x03C0) x 540(0x021C) )
                 board = board_t::board_M5Paper;
